Enhance test configurations and refactor imports
- Updated HealthCheckSpec and KeycloakContainerSpec with new timeout configurations and simplified imports.
- Refactored package names in Keycloak test syntax to align with the project structure.
- Upgraded Keycloak container version to `26.5.2`.
- Adjusted Docker build command structure and enabled Playwright browser installation in Test tasks.
